Update tailwind + cleanup queue

+++ src/lib/queues/cleanup.ts
@@ -1,21 +1,25 @@
+import { dev } from "$app/env"
 import { asyncExecShell, getEngine } from "$lib/common"
 import { prisma } from "$lib/database"
 
 export default async function () {
-    const destinationDockers = await prisma.destinationDocker.findMany()
-    for (const destinationDocker of destinationDockers) {
-        const host = getEngine(destinationDocker.engine)
-        try {
-            await asyncExecShell(`DOCKER_HOST=${host} docker container prune -f`)
-        } catch (error) {
-            //
-            console.log(error)
-        }
-        try {
-            await asyncExecShell(`DOCKER_HOST=${host} docker image prune --all -f`)
-        } catch (error) {
-            //
-            console.log(error)
+    if (!dev) {
+        const destinationDockers = await prisma.destinationDocker.findMany()
+        for (const destinationDocker of destinationDockers) {
+            const host = getEngine(destinationDocker.engine)
+            try {
+                await asyncExecShell(`DOCKER_HOST=${host} docker container prune -f`)
+            } catch (error) {
+                //
+                console.log(error)
+            }
+            try {
+                await asyncExecShell(`DOCKER_HOST=${host} docker image prune --all -f`)
+            } catch (error) {
+                //
+                console.log(error)
+            }
         }
     }
+
 }
